+++ zcash_client_sqlite/src/wallet/init/migrations.rs
@@ -1,3 +1,12 @@
+//! Migrations for the `zcash_client_sqlite` wallet database.
+//!
+//! The constants in this module cover all states of the migration DAG that have been
+//! exposed in a public crate release, in the order that crate users would have
+//! encountered them.
+//!
+//! Omitted versions had the same migration state as the first prior version that is
+//! included.
+
 mod add_account_birthdays;
 mod add_account_uuids;
 mod add_transaction_views;
@@ -173,7 +182,7 @@ pub(super) fn all_migrations<
             _rng: rng.clone(),
         }),
         Box::new(ensure_default_transparent_address::Migration {
-            params: params.clone(),
+            _params: params.clone(),
         }),
     ]
 }
@@ -190,20 +199,20 @@ const PUBLIC_MIGRATION_STATES: &[&[Uuid]] = &[
 ];
 
 /// Leaf migrations in the 0.4.0 release.
-const V_0_4_0: &[Uuid] = &[add_transaction_views::MIGRATION_ID];
+pub const V_0_4_0: &[Uuid] = &[add_transaction_views::MIGRATION_ID];
 
 /// Leaf migrations in the 0.6.0 release.
-const V_0_6_0: &[Uuid] = &[v_transactions_net::MIGRATION_ID];
+pub const V_0_6_0: &[Uuid] = &[v_transactions_net::MIGRATION_ID];
 
 /// Leaf migrations in the 0.8.0 release.
-const V_0_8_0: &[Uuid] = &[
+pub const V_0_8_0: &[Uuid] = &[
     nullifier_map::MIGRATION_ID,
     v_transactions_note_uniqueness::MIGRATION_ID,
     wallet_summaries::MIGRATION_ID,
 ];
 
 /// Leaf migrations in the 0.9.0 release.
-const V_0_9_0: &[Uuid] = &[
+pub const V_0_9_0: &[Uuid] = &[
     nullifier_map::MIGRATION_ID,
     receiving_key_scopes::MIGRATION_ID,
     v_transactions_note_uniqueness::MIGRATION_ID,
@@ -211,21 +220,21 @@ const V_0_9_0: &[Uuid] = &[
 ];
 
 /// Leaf migrations in the 0.10.0 release.
-const V_0_10_0: &[Uuid] = &[
+pub const V_0_10_0: &[Uuid] = &[
     nullifier_map::MIGRATION_ID,
     orchard_received_notes::MIGRATION_ID,
     orchard_shardtree::MIGRATION_ID,
 ];
 
 /// Leaf migrations in the 0.10.3 release.
-const V_0_10_3: &[Uuid] = &[
+pub const V_0_10_3: &[Uuid] = &[
     ensure_orchard_ua_receiver::MIGRATION_ID,
     nullifier_map::MIGRATION_ID,
     orchard_shardtree::MIGRATION_ID,
 ];
 
 /// Leaf migrations in the 0.11.0 release.
-const V_0_11_0: &[Uuid] = &[
+pub const V_0_11_0: &[Uuid] = &[
     ensure_orchard_ua_receiver::MIGRATION_ID,
     ephemeral_addresses::MIGRATION_ID,
     nullifier_map::MIGRATION_ID,
@@ -235,25 +244,25 @@ const V_0_11_0: &[Uuid] = &[
 ];
 
 /// Leaf migrations in the 0.11.1 release.
-const V_0_11_1: &[Uuid] = &[tx_retrieval_queue::MIGRATION_ID];
+pub const V_0_11_1: &[Uuid] = &[tx_retrieval_queue::MIGRATION_ID];
 
 /// Leaf migrations in the 0.11.2 release.
-const V_0_11_2: &[Uuid] = &[support_legacy_sqlite::MIGRATION_ID];
+pub const V_0_11_2: &[Uuid] = &[support_legacy_sqlite::MIGRATION_ID];
 
 /// Leaf migrations in the 0.12.0 release.
-const V_0_12_0: &[Uuid] = &[fix_broken_commitment_trees::MIGRATION_ID];
+pub const V_0_12_0: &[Uuid] = &[fix_broken_commitment_trees::MIGRATION_ID];
 
 /// Leaf migrations in the 0.13.0 release.
-const V_0_13_0: &[Uuid] = &[fix_bad_change_flagging::MIGRATION_ID];
+pub const V_0_13_0: &[Uuid] = &[fix_bad_change_flagging::MIGRATION_ID];
 
 /// Leaf migrations in the 0.14.0 release.
-const V_0_14_0: &[Uuid] = &[
+pub const V_0_14_0: &[Uuid] = &[
     fix_bad_change_flagging::MIGRATION_ID,
     add_account_uuids::MIGRATION_ID,
 ];
 
 /// Leaf migrations in the 0.15.0 release.
-const V_0_15_0: &[Uuid] = &[
+pub const V_0_15_0: &[Uuid] = &[
     fix_bad_change_flagging::MIGRATION_ID,
     v_transactions_additional_totals::MIGRATION_ID,
 ];
@@ -320,7 +329,7 @@ mod tests {
 
     use crate::{
         testing::db::{test_clock, test_rng},
-        wallet::init::init_wallet_db_internal,
+        wallet::init::WalletMigrator,
         WalletDb,
     };
 
@@ -338,12 +347,10 @@ mod tests {
 
         let seed = [0xab; 32];
         assert_matches!(
-            init_wallet_db_internal(
-                &mut db_data,
-                Some(Secret::new(seed.to_vec())),
-                migrations,
-                false
-            ),
+            WalletMigrator::new()
+                .with_seed(Secret::new(seed.to_vec()))
+                .ignore_seed_relevance()
+                .init_or_migrate_to(&mut db_data, migrations),
             Ok(_)
         );
     }
@@ -377,12 +384,10 @@ mod tests {
         let mut prev_leaves: &[Uuid] = &[];
         for migrations in super::PUBLIC_MIGRATION_STATES {
             assert_matches!(
-                init_wallet_db_internal(
-                    &mut db_data,
-                    Some(Secret::new(seed.clone())),
-                    migrations,
-                    false
-                ),
+                WalletMigrator::new()
+                    .with_seed(Secret::new(seed.clone()))
+                    .ignore_seed_relevance()
+                    .init_or_migrate_to(&mut db_data, migrations),
                 Ok(_)
             );
 
@@ -399,7 +404,10 @@ mod tests {
         // Now check that we can migrate from the last public release to the current
         // migration state in this branch.
         assert_matches!(
-            init_wallet_db_internal(&mut db_data, Some(Secret::new(seed)), &[], false),
+            WalletMigrator::new()
+                .with_seed(Secret::new(seed))
+                .ignore_seed_relevance()
+                .init_or_migrate(&mut db_data),
             Ok(_)
         );
         // We don't ensure that the migration state changed, because it may not have.